scene: use position independent scaling

Closes #3793, #3705, #3792, and #3139

See merge request wlroots/wlroots!5071

+++ test/bench_scene.c
@@ -104,7 +104,7 @@ static void bench_scene_node_at(struct wlr_scene *scene, struct tree_spec *spec)
 }
 
 static void noop_iterator(struct wlr_scene_buffer *buffer,
-		int sx, int sy, void *user_data) {
+		double sx, double sy, void *user_data) {
 	(void)buffer;
 	(void)sx;
 	(void)sy;
